+++ src/components/datatable/datatable.ts
@@ -1,0 +1,3 @@
+export const datatableSlotName = (columnName: string, prefix = 'item') => {
+  return prefix + '.' + columnName
+}