+++ src/services/api/apiAnyRequest.ts
@@ -58,6 +58,7 @@ export const apiAnyRequest = <T, R = T>(
         if (axiosErrorResponseHasForbiddenOperationData(err)) {
           return reject(new AnzuApiForbiddenOperationError(err, err))
         }
+        // todo catch another axios errors, for example timeout
         return reject(new AnzuFatalError(err))
       })
   })