+++ src/services/api/apiDeleteOne.ts
@@ -48,6 +48,7 @@ export const apiDeleteOne = <T, R = T>(
         if (axiosErrorResponseHasForbiddenOperationData(err)) {
           return reject(new AnzuApiForbiddenOperationError(err, err))
         }
+        // todo catch another axios errors, for example timeout
         return reject(new AnzuFatalError(err))
       })
   })