+++ src/services/api/apiFetchByIds.ts
@@ -63,6 +63,7 @@ export const apiFetchByIds = <T, R = T>(
         if (axiosErrorResponseHasForbiddenOperationData(err)) {
           return reject(new AnzuApiForbiddenOperationError(err, err))
         }
+        // todo catch another axios errors, for example timeout
         return reject(new AnzuFatalError(err))
       })
   })