+++ src/services/api/apiFetchOne.ts
@@ -47,6 +47,7 @@ export const apiFetchOne = <R>(
         if (axiosErrorResponseHasForbiddenOperationData(err)) {
           return reject(new AnzuApiForbiddenOperationError(err, err))
         }
+        // todo catch another axios errors, for example timeout
         return reject(new AnzuFatalError(err))
       })
   })